refactor(directory): tidy up!
js files except index.js is now inside of src/modules/

BREAKING CHANGE: directory tree changed!
